requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
fastapi==0.135.1
idna==3.11
inflect==7.5.0
more-itertools==10.8.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sqlacodegen==4.0.1
SQLAlchemy==2.0.47
starlette==0.52.1
typeguard==4.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
